speedy-murmurs-simulator/requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
ipykernel==5.1.1
ipyparallel==6.2.4
ipython==7.6.0
ipython-genutils==0.2.0
jedi==0.14.0
jupyter-client==5.2.4
jupyter-core==4.5.0
networkx==2.3
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.1
pyzmq==18.0.2
six==1.12.0
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
